+++ CHANGELOG.md
@@ -130,3 +130,8 @@
 ## 0.6.1
 
 - Completed EOS bug fix
+
+## 0.6.2
+
+- Improved doc for RuntimeError due to retraining FastText and BPEmb model in same directory.
+- Added error handling RuntimeError when retraining.